Create README.md
I was exploring different time series techniques on a relatively simple and clean dataset. This was a Kaggle competition where I was given 5 years of store-item sales data and asked to predict 3 months of sales for 50 different items at 10 other stores.
I decided that with deep learning it could work better and the pack I then used was xgboost.

Great competition was to help improve my forecasting skills.